+++ okta/resource_user_base_schema_test.go
@@ -3,9 +3,10 @@ package okta
 import (
 	"errors"
 	"fmt"
+	"testing"
+
 	"github.com/hashicorp/terraform/helper/resource"
 	"github.com/hashicorp/terraform/terraform"
-	"testing"
 
 	"github.com/hashicorp/terraform/helper/acctest"
 )
@@ -51,6 +52,7 @@ func TestAccOktaUserBaseSchemas(t *testing.T) {
 					resource.TestCheckResourceAttr(resourceName, "index", baseTestProp),
 					resource.TestCheckResourceAttr(resourceName, "title", "First name"),
 					resource.TestCheckResourceAttr(resourceName, "type", "string"),
+					resource.TestCheckResourceAttr(resourceName, "required", "true"),
 					resource.TestCheckResourceAttr(resourceName, "permissions", "READ_WRITE"),
 				),
 			},
@@ -60,7 +62,7 @@ func TestAccOktaUserBaseSchemas(t *testing.T) {
 
 func TestAccuserBaseSchemaImport(t *testing.T) {
 	ri := acctest.RandInt()
-	resourceName := buildResourceFQN(userBaseSchema, ri)
+	resourceName := fmt.Sprintf("%s.%s", userBaseSchema, baseTestProp)
 	mgr := newFixtureManager(userBaseSchema)
 	config := mgr.GetFixtures("basic.tf", ri, t)