+++ sdk/authorization_server_policies.go
@@ -1,4 +1,4 @@
-package okta
+package sdk
 
 import (
 	"fmt"
@@ -29,23 +29,23 @@ type (
 
 func (m *ApiSupplement) DeleteAuthorizationServerPolicy(authServerId, id string) (*okta.Response, error) {
 	url := fmt.Sprintf("/api/v1/authorizationServers/%s/policies/%s", authServerId, id)
-	req, err := m.requestExecutor.NewRequest("DELETE", url, nil)
+	req, err := m.RequestExecutor.NewRequest("DELETE", url, nil)
 	if err != nil {
 		return nil, err
 	}
 
-	return m.requestExecutor.Do(req, nil)
+	return m.RequestExecutor.Do(req, nil)
 }
 
 func (m *ApiSupplement) ListAuthorizationServerPolicies(authServerId string) ([]*AuthorizationServerPolicy, *okta.Response, error) {
 	url := fmt.Sprintf("/api/v1/authorizationServers/%s/policies", authServerId)
-	req, err := m.requestExecutor.NewRequest("GET", url, nil)
+	req, err := m.RequestExecutor.NewRequest("GET", url, nil)
 	if err != nil {
 		return nil, nil, err
 	}
 
 	var auth []*AuthorizationServerPolicy
-	resp, err := m.requestExecutor.Do(req, &auth)
+	resp, err := m.RequestExecutor.Do(req, &auth)
 	return auth, resp, err
 }
 
@@ -54,13 +54,13 @@ func (m *ApiSupplement) CreateAuthorizationServerPolicy(authServerId string, bod
 	if qp != nil {
 		url = url + qp.String()
 	}
-	req, err := m.requestExecutor.NewRequest("POST", url, body)
+	req, err := m.RequestExecutor.NewRequest("POST", url, body)
 	if err != nil {
 		return nil, nil, err
 	}
 
 	authorizationServer := body
-	resp, err := m.requestExecutor.Do(req, &authorizationServer)
+	resp, err := m.RequestExecutor.Do(req, &authorizationServer)
 	return &authorizationServer, resp, err
 }
 
@@ -69,13 +69,13 @@ func (m *ApiSupplement) UpdateAuthorizationServerPolicy(authServerId, id string,
 	if qp != nil {
 		url = url + qp.String()
 	}
-	req, err := m.requestExecutor.NewRequest("PUT", url, body)
+	req, err := m.RequestExecutor.NewRequest("PUT", url, body)
 	if err != nil {
 		return nil, nil, err
 	}
 
 	authorizationServer := body
-	resp, err := m.requestExecutor.Do(req, &authorizationServer)
+	resp, err := m.RequestExecutor.Do(req, &authorizationServer)
 	if err != nil {
 		return nil, resp, err
 	}
@@ -84,13 +84,13 @@ func (m *ApiSupplement) UpdateAuthorizationServerPolicy(authServerId, id string,
 
 func (m *ApiSupplement) GetAuthorizationServerPolicy(authServerId, id string, authorizationServerInstance AuthorizationServerPolicy) (*AuthorizationServerPolicy, *okta.Response, error) {
 	url := fmt.Sprintf("/api/v1/authorizationServers/%s/policies/%s", authServerId, id)
-	req, err := m.requestExecutor.NewRequest("GET", url, nil)
+	req, err := m.RequestExecutor.NewRequest("GET", url, nil)
 	if err != nil {
 		return nil, nil, err
 	}
 
 	authorizationServer := authorizationServerInstance
-	resp, err := m.requestExecutor.Do(req, &authorizationServer)
+	resp, err := m.RequestExecutor.Do(req, &authorizationServer)
 	if err != nil {
 		return nil, resp, err
 	}