+++ sdk/templates.go
@@ -1,4 +1,4 @@
-package okta
+package sdk
 
 import (
 	"fmt"
@@ -26,13 +26,13 @@ type (
 
 func (m *ApiSupplement) ListEmailTemplates() ([]*EmailTemplate, *okta.Response, error) {
 	url := "/api/v1/templates/emails"
-	req, err := m.requestExecutor.NewRequest("GET", url, nil)
+	req, err := m.RequestExecutor.NewRequest("GET", url, nil)
 	if err != nil {
 		return nil, nil, err
 	}
 
 	var auth []*EmailTemplate
-	resp, err := m.requestExecutor.Do(req, &auth)
+	resp, err := m.RequestExecutor.Do(req, &auth)
 	return auth, resp, err
 }
 
@@ -41,13 +41,13 @@ func (m *ApiSupplement) CreateEmailTemplate(id string, body EmailTemplate, qp *q
 	if qp != nil {
 		url += qp.String()
 	}
-	req, err := m.requestExecutor.NewRequest("POST", url, body)
+	req, err := m.RequestExecutor.NewRequest("POST", url, body)
 	if err != nil {
 		return nil, nil, err
 	}
 
 	temp := &EmailTemplate{}
-	resp, err := m.requestExecutor.Do(req, temp)
+	resp, err := m.RequestExecutor.Do(req, temp)
 	return temp, resp, err
 }
 
@@ -56,33 +56,33 @@ func (m *ApiSupplement) UpdateEmailTemplate(id string, body EmailTemplate, qp *q
 	if qp != nil {
 		url += qp.String()
 	}
-	req, err := m.requestExecutor.NewRequest("PUT", url, body)
+	req, err := m.RequestExecutor.NewRequest("PUT", url, body)
 	if err != nil {
 		return nil, nil, err
 	}
 
 	temp := &EmailTemplate{}
-	resp, err := m.requestExecutor.Do(req, temp)
+	resp, err := m.RequestExecutor.Do(req, temp)
 	return temp, resp, err
 }
 
 func (m *ApiSupplement) GetEmailTemplate(id string) (*EmailTemplate, *okta.Response, error) {
 	url := fmt.Sprintf("/api/v1/templates/emails/%s", id)
-	req, err := m.requestExecutor.NewRequest("GET", url, nil)
+	req, err := m.RequestExecutor.NewRequest("GET", url, nil)
 	if err != nil {
 		return nil, nil, err
 	}
 	temp := &EmailTemplate{}
-	resp, err := m.requestExecutor.Do(req, temp)
+	resp, err := m.RequestExecutor.Do(req, temp)
 	return temp, resp, err
 }
 
 func (m *ApiSupplement) DeleteEmailTemplate(id string) (*okta.Response, error) {
 	url := fmt.Sprintf("/api/v1/templates/emails/%s", id)
-	req, err := m.requestExecutor.NewRequest("DELETE", url, nil)
+	req, err := m.RequestExecutor.NewRequest("DELETE", url, nil)
 	if err != nil {
 		return nil, err
 	}
-	resp, err := m.requestExecutor.Do(req, nil)
+	resp, err := m.RequestExecutor.Do(req, nil)
 	return resp, err
 }